requirements.txt
aws-cdk-lib==2.139.0
constructs>=10.0.0,<11.0.0
jproperties>=2.1.1
